Aggiunto il @PathVariable per passare lo username in tutte le funzioni che necessitano di identificare uno user presente nel db.

+++ WineAdvisor/src/main/java/com/wineadvisor/wineadvisor/controller/UserController.java
@@ -2,9 +2,11 @@ package com.wineadvisor.wineadvisor.controller;
 
 import java.net.URI;
 
+import org.apache.coyote.BadRequestException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -28,6 +30,10 @@ public class UserController {
     @PostMapping
     public ResponseEntity<?> addUser(@RequestBody User newUser) {
         try {
+            if (newUser == null) {
+                throw new BadRequestException("User cannot be null");
+            }
+            
             return ResponseEntity.created(URI.create("/api/user/" + newUser.getLogin().getUsername())).body(userService.addUser(newUser));
         } catch (Exception e) {
             System.err.println("--- ERR: " + e.getMessage());
@@ -46,17 +52,7 @@ public class UserController {
         }
     }
 
-    @GetMapping("/username")
-    public ResponseEntity<?> getUserByUsername(@RequestParam String username) {
-        try {
-            return ResponseEntity.ok().body(userService.getUserByUsername(username));
-        } catch (Exception e) {
-            System.err.println("--- ERR: " + e.getMessage());
-            return ResponseEntity.badRequest().body(e.getMessage());
-        }
-    }
-
-    @GetMapping("/firstname")
+    @GetMapping
     public ResponseEntity<?> getUsersByFirstName(@RequestParam String firstName) {
         try {
             return ResponseEntity.ok().body(userService.getUsersByFirstName(firstName));
@@ -66,7 +62,7 @@ public class UserController {
         }
     }
 
-    @GetMapping("/lastname")
+    @GetMapping
     public ResponseEntity<?> getUsersByLastName(@RequestParam String lastName) {
         try {
             return ResponseEntity.ok().body(userService.getUsersByLastName(lastName));
@@ -76,11 +72,33 @@ public class UserController {
         }
     }
 
+    @GetMapping("/{username}")
+    public ResponseEntity<?> getUserByUsername(@PathVariable String username) {
+        try {
+            if (username == null || username.isEmpty()) {
+                throw new BadRequestException("Username cannot be null or empty");
+            }
+
+            return ResponseEntity.ok().body(userService.getUserByUsername(username));
+        } catch (Exception e) {
+            System.err.println("--- ERR: " + e.getMessage());
+            return ResponseEntity.badRequest().body(e.getMessage());
+        }
+    }
+
     
     ////////////// PUT //////////////
-    @PutMapping
-    public ResponseEntity<?> updateUser(@RequestBody User user) {
+    @PutMapping("/{username}")
+    public ResponseEntity<?> updateUser(@PathVariable String username, @RequestBody User user) {
         try {
+            if (username == null || username.isEmpty()) {
+                throw new BadRequestException("Username cannot be null or empty");
+            }
+            if (user == null) {
+                throw new BadRequestException("User cannot be null");
+            }
+
+            user.getLogin().setUsername(username);
             return ResponseEntity.ok().body(userService.updateUser(user));
         } catch (Exception e) {
             System.err.println("--- ERR: " + e.getMessage());
@@ -88,27 +106,39 @@ public class UserController {
         }
     }
 
-    @PutMapping("/password/update")
-    public ResponseEntity<?> updateUserPassword(@RequestBody UpdatePasswordRequest body) {
+    @PutMapping("{username}/password/update")
+    public ResponseEntity<?> updateUserPassword(@PathVariable String username, @RequestBody UpdatePasswordRequest updatePasswordRequest) {
         try {
-            String username = body.getUsername();
-            String oldPass = body.getOldPass();
-            String newPass = body.getNewPass();
-            String confirmPass = body.getConfirmPass();
-            
-            return ResponseEntity.ok().body(userService.updateUserPassword(username, oldPass, newPass, confirmPass));
+            if (username == null || username.isEmpty()) {
+                throw new BadRequestException("Username cannot be null or empty");
+            }
+            if (updatePasswordRequest.getOldPass() == null || updatePasswordRequest.getOldPass().isEmpty()) {
+                throw new BadRequestException("Old password cannot be null or empty");
+            }
+            if (updatePasswordRequest.getNewPass() == null || updatePasswordRequest.getNewPass().isEmpty()) {
+                throw new BadRequestException("New password cannot be null or empty");
+            }
+            if (updatePasswordRequest.getConfirmPass() == null || updatePasswordRequest.getConfirmPass().isEmpty()) {
+                throw new BadRequestException("Confirm password cannot be null or empty");
+            }
+
+            return ResponseEntity.ok().body(userService.updateUserPassword(username, updatePasswordRequest.getOldPass(), updatePasswordRequest.getNewPass(), updatePasswordRequest.getConfirmPass()));
         } catch (Exception e) {
             System.err.println("--- ERR: " + e.getMessage());
             e.printStackTrace();
             return ResponseEntity.badRequest().body(e.getMessage());
         }
     }
-    
+
 
     ////////////// DELETE //////////////
     @DeleteMapping
-    public ResponseEntity<?> deleteUser(@RequestParam String username) {
+    public ResponseEntity<?> deleteUser(@PathVariable String username) {
         try {
+            if (username == null || username.isEmpty()) {
+                throw new BadRequestException("Username cannot be null or empty");
+            }
+
             userService.deleteUser(username);
             return ResponseEntity.ok().body("Utente eliminato correttamente");
         } catch (Exception e) {
